rename actlogutils -> parser
