+++ act/nari/ext/act/__init__.py
@@ -6,7 +6,7 @@ from nari.types.event import Event
 from nari.types.event.version import SemanticVersion, Version
 from nari.util.exceptions import EventNotFound
 
-from nari.ext.act.actlogutils import ID_MAPPINGS, ActEventType, ActLogChecksumType, date_from_act_timestamp, validate_checksum
+from nari.ext.act.parser import ID_MAPPINGS, ActEventType, ActLogChecksumType, date_from_act_timestamp, validate_checksum
 from nari.ext.act.exceptions import InvalidActChecksum
 
 # The last version of ACT to use MD5 checksums
