+++ act/test/test_updatehpmp_parsing.py
@@ -1,7 +1,7 @@
 import unittest
 
 from nari.types.event.updatehpmp import UpdateHpMp
-from nari.ext.act.actlogutils.updatehpmp import updatehpmp_from_logline
+from nari.ext.act.parser.updatehpmp import updatehpmp_from_logline
 
 
 class TestUpdateHPMPFromACTLogLine(unittest.TestCase):
